Fix 64-bit incompatibility that broke TS_ASSERT_SAME_DATA
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7158 3db68df2-c116-0410-a063-a993310a9797

+++ libraries/cxxtest/include/cxxtest/ErrorFormatter.h
@@ -129,7 +129,7 @@ namespace CxxTest
         void failedAssertSameData( const char *file, unsigned line,
                                    const char *xStr, const char *yStr,
                                    const char *sizeStr, const void *x,
-                                   const void *y, unsigned size )
+                                   const void *y, size_t size )
         {
             stop( file, line ) << "Error: Expected " << sizeStr << " (" << size << ") bytes to be equal at (" <<
                 xStr << ") and (" << yStr << "), found:" << endl;
